Break out client-package-specific tests into separate GeophiresXClientTestCase
